Creation of Service interfaces and use them instead in controllers, constructor injection implementation in controllers
