+++ src/main/java/com/persado/assignment/project/controller/HomeController.java
@@ -1,7 +1,6 @@
 package com.persado.assignment.project.controller;
 
 import java.time.LocalDateTime;
-
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
